plotting data

+++ scripts/rsl_rl/play_and_plot.py
@@ -19,7 +19,7 @@ import cli_args  # isort: skip
 parser = argparse.ArgumentParser(description="Train an RL agent with RSL-RL.")
 parser.add_argument("--video", action="store_true", default=False, help="Record videos during training.")
 parser.add_argument("--video_length", type=int, default=200, help="Length of the recorded video (in steps).")
-parser.add_argument("--mass", type=int, default=13, help="Mass of torso")
+parser.add_argument("--mass", type=float, default=13.0, help="Mass of torso")
 
 parser.add_argument(
     "--disable_fabric", action="store_true", default=False, help="Disable fabric and use USD I/O operations."
